事务

+++ src/main/java/com/it/community/service/QuestionService.java
@@ -65,7 +65,7 @@ public class QuestionService {
     }
 
 
-    public PageInfo list(Integer userId, Integer page, Integer size) {
+    public PageInfo list(Long userId, Integer page, Integer size) {
         PageInfo pageInfo = new PageInfo();
 
         Integer totalPage;//总页数
@@ -105,7 +105,7 @@ public class QuestionService {
         return pageInfo;
     }
 
-    public QuestionDTO getById(Integer id) {
+    public QuestionDTO getById(Long id) {
         Question question = questionMapper.selectByPrimaryKey(id);
         if (question==null){
             throw new CustomizeException(CustomizeErrorCode.QUESTION_NOT_FOUND);
@@ -139,7 +139,7 @@ public class QuestionService {
         }
     }
 
-    public void incView(Integer id) {
+    public void incView(Long id) {
         Question question = new Question();
         question.setId(id);
         question.setViewCount(1);
